backend/requirements.txt
Django==3.2.16
djangorestframework==3.14.0
djoser==2.1.0
Pillow>=10.0.0
gunicorn==20.1.0
psycopg2-binary
django-filter==22.1
python-dotenv==0.21.0
drf-extra-fields==3.4.0
drf-spectacular==0.26.5
